requirements.txt
asgiref==3.8.1
attrs==23.2.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
defusedxml==0.7.1
dj-rest-auth==6.0.0
Django==4.0
django-allauth==0.61.1
django-cors-headers==4.3.1
djangorestframework==3.15.1
drf-spectacular==0.27.2
idna==3.7
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
oauthlib==3.2.2
pycparser==2.22
PyJWT==2.8.0
python-environ==0.4.54
python3-openid==3.2.0
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.18.1
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
